changing appearance done;

+++ Backend/src/main/java/pl/HomeworkJustClick/Backend/Services/GroupStudentServiceImplement.java
@@ -1,30 +1,27 @@
 package pl.HomeworkJustClick.Backend.Services;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.HomeworkJustClick.Backend.Entities.GroupStudent;
-import pl.HomeworkJustClick.Backend.Entities.GroupTeacher;
 import pl.HomeworkJustClick.Backend.Repositories.GroupRepository;
 import pl.HomeworkJustClick.Backend.Repositories.GroupStudentRepository;
 import pl.HomeworkJustClick.Backend.Repositories.GroupTeacherRepository;
 import pl.HomeworkJustClick.Backend.Repositories.UserRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class GroupStudentServiceImplement implements GroupStudentService{
 
-    @Autowired
-    GroupStudentRepository groupStudentRepository;
+    private final GroupStudentRepository groupStudentRepository;
 
-    @Autowired
-    GroupRepository groupRepository;
+    private final GroupRepository groupRepository;
 
-    @Autowired
-    GroupTeacherRepository groupTeacherRepository;
+    private final GroupTeacherRepository groupTeacherRepository;
 
-    @Autowired
-    UserRepository userRepository;
+    private final UserRepository userRepository;
 
     @Override
     public List<GroupStudent> getAll() {
@@ -32,12 +29,8 @@ public class GroupStudentServiceImplement implements GroupStudentService{
     }
 
     @Override
-    public GroupStudent getById(int id) {
-        if (groupStudentRepository.findById(id).isPresent()) {
-            return groupStudentRepository.findById(id).get();
-        } else {
-            return null;
-        }
+    public Optional<GroupStudent> getById(int id) {
+        return groupStudentRepository.findById(id);
     }
 
     @Override
@@ -48,10 +41,10 @@ public class GroupStudentServiceImplement implements GroupStudentService{
 
     @Override
     public Boolean delete(int id) {
-        try {
+        if(groupStudentRepository.existsById(id)) {
             groupStudentRepository.deleteById(id);
             return true;
-        } catch (IllegalArgumentException e) {
+        } else {
             return false;
         }
     }
@@ -64,15 +57,18 @@ public class GroupStudentServiceImplement implements GroupStudentService{
             groupStudentRepository.save(groupStudent);
             return true;
         } else {
-            return null;
+            return false;
         }
     }
 
     @Override
     public Boolean addStudentToGroup (int group_id, int student_id) {
-        int groupTeacherCheck = groupTeacherRepository.getGroupTeacherByTeacherAndGroup(student_id, group_id);
-        int groupStudentCheck = groupStudentRepository.getGroupStudentByStudentAndGroup(student_id, group_id);
-        if (groupTeacherCheck == 0 && groupStudentCheck == 0 && groupRepository.findById(group_id).isPresent() && userRepository.findById(student_id).isPresent()) {
+        int groupTeacherCheck = groupTeacherRepository.checkForTeacherInGroup(student_id, group_id);
+        int groupStudentCheck = groupStudentRepository.checkForStudentInGroup(student_id, group_id);
+        if(groupTeacherCheck != 0 || groupStudentCheck != 0){
+            return null;
+        }
+        else if (groupRepository.findById(group_id).isPresent() && userRepository.findById(student_id).isPresent()) {
             GroupStudent groupStudent = new GroupStudent(groupRepository.findById(group_id).get(), userRepository.findById(student_id).get(), "");
             groupStudentRepository.save(groupStudent);
             return true;
@@ -83,12 +79,18 @@ public class GroupStudentServiceImplement implements GroupStudentService{
 
     @Override
     public Boolean deleteStudentFromGroup (int group_id, int student_id) {
-        if (groupStudentRepository.getGroupStudentByStudentAndGroup(student_id, group_id) != 0){
+        if (groupStudentRepository.checkForStudentInGroup(student_id, group_id) != 0){
             GroupStudent groupStudent = groupStudentRepository.getGroupStudentObjectByStudentAndGroup(student_id, group_id);
             groupStudentRepository.deleteById(groupStudent.getId());
             return true;
         } else {
             return false;
         }
+    }
+
+    @Override
+    public Boolean checkForStudentInGroup(int student_id, int group_id) {
+        int groupStudentCheck = groupStudentRepository.checkForStudentInGroup(student_id, group_id);
+        return groupStudentCheck != 0;
     }
 }
