+++ Backend/src/main/java/pl/HomeworkJustClick/Backend/Services/UserServiceImplement.java
@@ -1,22 +1,21 @@
 package pl.HomeworkJustClick.Backend.Services;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pl.HomeworkJustClick.Backend.Entities.GroupTeacher;
 import pl.HomeworkJustClick.Backend.Entities.User;
 import pl.HomeworkJustClick.Backend.Repositories.GroupTeacherRepository;
 import pl.HomeworkJustClick.Backend.Repositories.UserRepository;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class UserServiceImplement implements UserService{
-    @Autowired
-    UserRepository userRepository;
 
-    @Autowired
-    GroupTeacherRepository groupTeacherRepository;
+    private final UserRepository userRepository;
+
+    private final GroupTeacherRepository groupTeacherRepository;
 
     @Override
     public List<User> getAll() {
@@ -24,12 +23,8 @@ public class UserServiceImplement implements UserService{
     }
 
     @Override
-    public User getById(int id) {
-        if (userRepository.findById(id).isPresent()) {
-            return userRepository.findById(id).get();
-        } else {
-            return null;
-        }
+    public Optional<User> getById(int id) {
+        return userRepository.findById(id);
     }
 
     @Override
@@ -42,16 +37,18 @@ public class UserServiceImplement implements UserService{
     public Boolean update(int id, User updatedUser) {
         if (userRepository.findById(id).isPresent()) {
             User user = userRepository.getReferenceById(id);
-            if (updatedUser.getFirstname() != null) {
+            if (!updatedUser.getFirstname().isEmpty() && !updatedUser.getFirstname().equals(user.getFirstname())) {
                 user.setFirstname(updatedUser.getFirstname());
             }
-            if (updatedUser.getLastname() != null) {
+            if (!updatedUser.getLastname().isEmpty() && !updatedUser.getLastname().equals(user.getLastname())) {
                 user.setLastname(updatedUser.getLastname());
             }
             if (updatedUser.getIndex() != user.getIndex()) {
                 user.setIndex(updatedUser.getIndex());
             }
-
+            if (updatedUser.getColor() != user.getColor()) {
+                user.setColor(updatedUser.getColor());
+            }
             try {
                 userRepository.save(user);
                 return true;
@@ -59,16 +56,16 @@ public class UserServiceImplement implements UserService{
                 return false;
             }
         } else {
-            return false;
+            return null;
         }
     }
 
     @Override
     public Boolean delete(int id) {
-        try {
+        if(userRepository.existsById(id)) {
             userRepository.deleteById(id);
             return true;
-        } catch (IllegalArgumentException e) {
+        } else  {
             return false;
         }
     }
@@ -81,17 +78,29 @@ public class UserServiceImplement implements UserService{
             userRepository.save(user);
             return true;
         } else {
-            return null;
+            return false;
+        }
+    }
+
+    @Override
+    public Boolean changeColorById(int id, int color) {
+        if (userRepository.findById(id).isPresent()) {
+            User user = userRepository.findById(id).get();
+            user.setColor(color);
+            userRepository.save(user);
+            return true;
+        } else {
+            return false;
         }
     }
 
     @Override
     public List<User> getTeachersByGroup(int group_id) {
-        return userRepository.getGroupTeachersByGroup(group_id);
+        return userRepository.getTeachersByGroupId(group_id);
     }
 
     @Override
     public List<User> getStudentsByGroup(int group_id) {
-        return userRepository.getGroupStudentsByGroup(group_id);
+        return userRepository.getStudentsByGroupId(group_id);
     }
 }
